requirements.txt
asgiref==3.7.2
Django==4.1.10
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
inflection==0.5.1
Pillow==10.0.0
pip==22.3.1
python-dotenv==0.21.0
pytz==2023.3
setuptools==65.5.0
sqlparse==0.4.4
tzdata==2023.3
django-cors-headers==4.2.0